ensure that all mapper/reducer outputs are tab separated
Streaming Hadoop ensures that every line emitted by a mapper or a reducer
includes a separator character to delimit the key and value. The default
separator character is a \t. Hadoop interprets any line without a
separator to be a key with a null value and appends a separator to the
end of the line. This commit changes Hasta to mimic this behavior.